requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt
psycopg2
